# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/process-file":{"post":{"summary":"Process File Endpoint","description":"Process a file and send the results to a callback URL.\n\nThe callback URL will receive a POST request with one of these JSON payloads:\n\nOn success:\n```json\n{\n    \"status\": \"success\",\n    \"hl7_fhir_data\": \"<FHIR resource JSON>\",\n    \"raw_text\": \"<extracted text>\",\n    \"mortal_readable\": \"<human readable text>\"\n}\n```\n\nOn error:\n```json\n{\n    \"status\": \"error\", \n    \"error\": \"<error message>\"\n}\n    ```","operationId":"process_file_endpoint_process_file_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileUrlRequest"}}},"required":true},"responses":{"201":{"description":"File processing started","content":{"application/json":{"schema":{"type":"string","enum":["OK"],"const":"OK","title":"Response Process File Endpoint Process File Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/suggestions":{"post":{"summary":"Get Suggestions","description":"Generate AI-assisted medical suggestions based on chat history.","operationId":"get_suggestions_suggestions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuggestionsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Suggestion"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatMessage":{"properties":{"is_organisation_message":{"type":"boolean","title":"Is Organisation Message"},"content":{"type":"string","title":"Content"}},"type":"object","required":["is_organisation_message","content"],"title":"ChatMessage"},"Document":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"metadata":{"type":"object","title":"Metadata"},"page_content":{"type":"string","title":"Page Content"},"type":{"type":"string","enum":["Document"],"const":"Document","title":"Type","default":"Document"}},"type":"object","required":["page_content"],"title":"Document","description":"Class for storing a piece of text and associated metadata.\n\nExample:\n\n    .. code-block:: python\n\n        from langchain_core.documents import Document\n\n        document = Document(\n            page_content=\"Hello, world!\",\n            metadata={\"source\": \"https://example.com\"}\n        )"},"FileUrlRequest":{"properties":{"url":{"type":"string","title":"Url"},"callback_url":{"type":"string","title":"Callback Url"},"patient_id":{"type":"string","title":"Patient Id"}},"type":"object","required":["url","callback_url","patient_id"],"title":"FileUrlRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Suggestion":{"properties":{"suggestion":{"type":"string","title":"Suggestion"},"sources":{"items":{"$ref":"#/components/schemas/Document"},"type":"array","title":"Sources"}},"type":"object","required":["suggestion","sources"],"title":"Suggestion"},"SuggestionsRequest":{"properties":{"patient_id":{"type":"string","title":"Patient Id"},"chat_history":{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array","title":"Chat History"}},"type":"object","required":["patient_id","chat_history"],"title":"SuggestionsRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}